karma_index/karma_rag_bot_ui-1/requirements.txt
google-generativeai
langchain
langchain-community
tkinter
PyQt5